+++ syncopy/preproc/compRoutines.py
@@ -10,7 +10,7 @@ import scipy.signal as sci
 from inspect import signature
 
 # syncopy imports
-from syncopy.shared.computational_routine import ComputationalRoutine
+from syncopy.shared.computational_routine import ComputationalRoutine, propagate_metadata
 from syncopy.shared.const_def import spectralConversions, spectralDTypes
 from syncopy.shared.kwarg_decorators import process_io
 
@@ -155,18 +155,7 @@ class SincFiltering(ComputationalRoutine):
 
     def process_metadata(self, data, out):
 
-        # Some index gymnastics to get trial begin/end "samples"
-        if data.selection is not None:
-            chanSec = data.selection.channel
-            trl = data.selection.trialdefinition
-        else:
-            chanSec = slice(None)
-            trl = data.trialdefinition
-
-        out.trialdefinition = trl
-
-        out.samplerate = data.samplerate
-        out.channel = np.array(data.channel[chanSec])
+        propagate_metadata(data, out)
 
 
 @process_io
@@ -289,18 +278,7 @@ class ButFiltering(ComputationalRoutine):
 
     def process_metadata(self, data, out):
 
-        # Some index gymnastics to get trial begin/end "samples"
-        if data.selection is not None:
-            chanSec = data.selection.channel
-            trl = data.selection.trialdefinition
-        else:
-            chanSec = slice(None)
-            trl = data.trialdefinition
-
-        out.trialdefinition = trl
-
-        out.samplerate = data.samplerate
-        out.channel = np.array(data.channel[chanSec])
+        propagate_metadata(data, out)
 
 
 @process_io
@@ -362,18 +340,7 @@ class Rectify(ComputationalRoutine):
 
     def process_metadata(self, data, out):
 
-        # Some index gymnastics to get trial begin/end "samples"
-        if data.selection is not None:
-            chanSec = data.selection.channel
-            trl = data.selection.trialdefinition
-        else:
-            chanSec = slice(None)
-            trl = data.trialdefinition
-
-        out.trialdefinition = trl
-
-        out.samplerate = data.samplerate
-        out.channel = np.array(data.channel[chanSec])
+        propagate_metadata(data, out)
 
 
 @process_io
@@ -451,19 +418,7 @@ class Hilbert(ComputationalRoutine):
 
     def process_metadata(self, data, out):
 
-        # Some index gymnastics to get trial begin/end "samples"
-        if data.selection is not None:
-            chanSec = data.selection.channel
-            trl = data.selection.trialdefinition
-
-        else:
-            chanSec = slice(None)
-            trl = data.trialdefinition
-
-        out.trialdefinition = trl
-
-        out.samplerate = data.samplerate
-        out.channel = np.array(data.channel[chanSec])
+        propagate_metadata(data, out)
 
 
 @process_io
@@ -475,7 +430,7 @@ def downsample_cF(dat,
                   noCompute=False
                   ):
     """
-    Provides basic downsampling of signals. The `new_samplerate` should be 
+    Provides basic downsampling of signals. The `new_samplerate` should be
     an integer division of the original `samplerate`.
 
     dat : (N, K) :class:`numpy.ndarray`
@@ -546,14 +501,15 @@ class Downsample(ComputationalRoutine):
     def process_metadata(self, data, out):
 
         # we need to re-calculate the downsampling factor
-        factor = int(data.samplerate // self.cfg['new_samplerate'])
+        # that it actually is an 1 / integer gets checked in the frontend
+        factor = self.cfg['new_samplerate'] / data.samplerate
 
         if data.selection is not None:
             chanSec = data.selection.channel
-            trl = data.selection.trialdefinition // factor
+            trl = _resampling_trl_definition(data.selection.trialdefinition, factor)
         else:
             chanSec = slice(None)
-            trl = data.trialdefinition // factor
+            trl = _resampling_trl_definition(data.trialdefinition, factor)
 
         out.trialdefinition = trl
         # now set new samplerate
@@ -660,17 +616,228 @@ class Resample(ComputationalRoutine):
 
         # we need to re-calculate the resampling factor
         factor = self.cfg['new_samplerate'] / data.samplerate
-        trafo_trl = lambda trldef: np.ceil(trldef * factor)
+        trafo_trl = _resampling_trl_definition
 
         if data.selection is not None:
             chanSec = data.selection.channel
-            trl = trafo_trl(data.selection.trialdefinition)
+            trl = trafo_trl(data.selection.trialdefinition, factor)
         else:
             chanSec = slice(None)
-            trl = trafo_trl(data.trialdefinition)
+            trl = trafo_trl(data.trialdefinition, factor)
 
         out.trialdefinition = trl
 
         # now set new samplerate
         out.samplerate = self.cfg['new_samplerate']
         out.channel = np.array(data.channel[chanSec])
+
+
+@process_io
+def detrending_cF(dat, polyremoval=None, timeAxis=0, noCompute=False, chunkShape=None):
+
+    """
+    Simple cF to wire SciPy's `detrend` to our CRs,
+    supported are constant and linear detrending
+
+    Parameters
+    ----------
+    dat : (N, K) :class:`numpy.ndarray`
+        Uniformly sampled multi-channel time-series data
+        The 1st dimension is interpreted as the time axis,
+        columns represent individual channels.
+        Dimensions can be transposed to `(K, N)` with the `timeAxis` parameter
+    polyremoval : {0, 1} or None
+        Order of polynomial used for de-trending data in the time domain.
+        A value of 0 corresponds to subtracting the mean
+        ("de-meaning"), ``polyremoval = 1`` removes linear trends (subtracting the
+        least squares fit of a linear polynomial).
+    timeAxis : int, optional
+        Index of running time axis in `dat` (0 or 1)
+    noCompute : bool
+        Preprocessing flag. If `True`, do not perform actual calculation but
+        instead return expected shape and :class:`numpy.dtype` of output
+        array.
+
+    Returns
+    -------
+    detrended : (N, K) :class:`~numpy.ndarray`
+        The detrended signals
+
+    Notes
+    -----
+    This method is intended to be used as
+    :meth:`~syncopy.shared.computational_routine.ComputationalRoutine.computeFunction`
+    inside a :class:`~syncopy.shared.computational_routine.ComputationalRoutine`.
+    Thus, input parameters are presumed to be forwarded from a parent metafunction.
+    Consequently, this function does **not** perform any error checking and operates
+    under the assumption that all inputs have been externally validated and cross-checked.
+    """
+
+    # should be captured in frontend
+    if polyremoval is None:
+        return dat
+
+    # Re-arrange array if necessary and get dimensional information
+    if timeAxis != 0:
+        dat = dat.T       # does not copy but creates view of `dat`
+    else:
+        dat = dat
+
+    # detrending does not change the shape
+    outShape = dat.shape
+    if noCompute:
+        return outShape, np.float32
+
+    # detrend
+    if polyremoval == 0:
+        dat = sci.detrend(dat, type='constant', axis=0, overwrite_data=True)
+    elif polyremoval == 1:
+        dat = sci.detrend(dat, type='linear', axis=0, overwrite_data=True)
+
+    # renaming
+    detrended = dat
+    return detrended
+
+
+class Detrending(ComputationalRoutine):
+
+    """
+    Compute class that performs constant or linear detrending
+    of :class:`~syncopy.AnalogData` objects
+
+    Sub-class of :class:`~syncopy.shared.computational_routine.ComputationalRoutine`,
+    see :doc:`/developer/compute_kernels` for technical details on Syncopy's compute
+    classes and metafunctions.
+
+    See also
+    --------
+    syncopy.preprocessing : parent metafunction
+    """
+
+    computeFunction = staticmethod(detrending_cF)
+
+    # 1st argument,the data, gets omitted
+    valid_kws = list(signature(detrending_cF).parameters.keys())[1:]
+
+    def process_metadata(self, data, out):
+
+        propagate_metadata(data, out)
+
+
+@process_io
+def standardize_cF(dat, polyremoval=None, timeAxis=0, noCompute=False, chunkShape=None):
+
+    """
+    Yet another simple cF to z-score ('standardize') signals:
+    subtracting the mean and normalize by standard deviation.
+
+    Parameters
+    ----------
+    dat : (N, K) :class:`numpy.ndarray`
+        Uniformly sampled multi-channel time-series data
+        The 1st dimension is interpreted as the time axis,
+        columns represent individual channels.
+        Dimensions can be transposed to `(K, N)` with the `timeAxis` parameter
+    polyremoval : {0, 1} or None
+        Order of polynomial used for de-trending data in the time domain prior
+        to z-scoring. A value of 0 corresponds to subtracting the mean
+        ("de-meaning"), ``polyremoval = 1`` removes linear trends (subtracting the
+        least squares fit of a linear polynomial).
+    timeAxis : int, optional
+        Index of running time axis in `dat` (0 or 1)
+    noCompute : bool
+        Preprocessing flag. If `True`, do not perform actual calculation but
+        instead return expected shape and :class:`numpy.dtype` of output
+        array.
+
+    Returns
+    -------
+    standardized : (N, K) :class:`~numpy.ndarray`
+        The standardized signals
+
+    Notes
+    -----
+    This method is intended to be used as
+    :meth:`~syncopy.shared.computational_routine.ComputationalRoutine.computeFunction`
+    inside a :class:`~syncopy.shared.computational_routine.ComputationalRoutine`.
+    Thus, input parameters are presumed to be forwarded from a parent metafunction.
+    Consequently, this function does **not** perform any error checking and operates
+    under the assumption that all inputs have been externally validated and cross-checked.
+    """
+
+    # Re-arrange array if necessary and get dimensional information
+    if timeAxis != 0:
+        dat = dat.T       # does not copy but creates view of `dat`
+    else:
+        dat = dat
+
+    # cF does not change the shape
+    outShape = dat.shape
+    if noCompute:
+        return outShape, np.float32
+
+    # detrend
+    if polyremoval == 0:
+        dat = sci.detrend(dat, type='constant', axis=0, overwrite_data=True)
+    elif polyremoval == 1:
+        dat = sci.detrend(dat, type='linear', axis=0, overwrite_data=True)
+
+    # standardize
+    dat = (dat - np.mean(dat, axis=0)) / np.std(dat, axis=0)
+
+    # renaming
+    standardized = dat
+    return standardized
+
+
+class Standardize(ComputationalRoutine):
+
+    """
+    Compute class that performs standardizing
+    of :class:`~syncopy.AnalogData` objects
+
+    Sub-class of :class:`~syncopy.shared.computational_routine.ComputationalRoutine`,
+    see :doc:`/developer/compute_kernels` for technical details on Syncopy's compute
+    classes and metafunctions.
+
+    See also
+    --------
+    syncopy.preprocessing : parent metafunction
+    """
+
+    computeFunction = staticmethod(standardize_cF)
+
+    # 1st argument,the data, gets omitted
+    valid_kws = list(signature(standardize_cF).parameters.keys())[1:]
+
+    def process_metadata(self, data, out):
+
+        propagate_metadata(data, out)
+
+
+def _resampling_trl_definition(orig_trl, factor):
+
+    """
+    Construct new trialdefinition from original
+    trialdefinition and the resampling factor
+    """
+
+    # start from input trial lengths and scale
+    # and ceil them to arrive at new trial lengths
+    # important is 1st diff then ceil..
+    sinfo = orig_trl[:, :2]
+    trl_len = np.ceil(np.diff(sinfo * factor, axis=1)).squeeze()
+
+    # use ceil again to define new trial start
+    # and offset samples
+    trl_scaled = np.ceil(orig_trl * factor)
+    trl_starts = trl_scaled[:, 0]
+    offsets = trl_scaled[:, 2]
+
+    # now add new trl_len to get new trial ends
+    trl_ends = trl_starts + trl_len
+
+    # finally stack everything back together
+    trldef = np.column_stack([trl_starts, trl_ends, offsets])
+
+    return trldef