+++ CHANGELOG.md
@@ -6,14 +6,15 @@ and this project follows [Semantic Versioning](https://semver.org/spec/v2.0.0.ht
 
 
 
-## [Unreleased]
+## [2022.08] - 2022-08-10
 
 ### NEW
 - Added down- and resampling algorithms for the new meta-function `resampledata`
 - Added FOOOF method as a post-processing option for the freqanalysis method mtmfft.
-- new global `spy.copy()` function which copies entire Syncopy objects on disk
+- Added `load_tdt` to import data from the TDT system, thanks to @kajal5888
 - Added `.info` attribute for all data classes to store auxiliary meta information
-
+- Added `zscore` normalization to `preprocessing`
+- new global `spy.copy()` function which copies entire Syncopy objects on disk
 
 ### CHANGED
 - the `out.cfg` attached to an analysis result now allows to replay all analysis methods
@@ -21,6 +22,7 @@ and this project follows [Semantic Versioning](https://semver.org/spec/v2.0.0.ht
 - `spy.cleanup` now has exposed `interactive` parameter
 - removed keyword `deep` from `copy()`, all our copies are in fact deep
 - demeaning after tapering for granger analysis
+- detrending is now possible without filtering in `preprocessing`
 
 ### FIXED
 - `out.cfg` global side-effects (sorry again @kajal5888)